+++ src/ska/contrib/django/ska/tests/test_constance_authentication_backend_.py
@@ -26,7 +26,7 @@ import factories
 
 from .helpers import log_info
 
-__title__ = 'ska.contrib.django.ska.tests.test_django_ska_constance'
+__title__ = 'ska.contrib.django.ska.tests.test_constance_authentication_backend'
 __author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
 __copyright__ = '2013-2019 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'