+++ src/ska/contrib/django/ska/tests/test_default_authentication_backend.py
@@ -22,7 +22,7 @@ import factories
 
 from .helpers import log_info
 
-__title__ = 'ska.contrib.django.ska.tests.test_django_ska'
+__title__ = 'ska.contrib.django.ska.tests.test_default_authentication_backend'
 __author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
 __copyright__ = '2013-2019 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'